# back-api/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: SmartHeart API
  description: API for EKG image analysis with GPT interpretation.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/requests/{id}:
    get:
      tags: [requests]
      summary: Get request with response and files
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/requests:
    get:
      tags: [requests]
      summary: List user's requests (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Request" }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Request:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        text_query: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        files:
          type: array
          items: { $ref: "#/components/schemas/File" }
        response: { $ref: "#/components/schemas/Response" }
    File:
      type: object
      properties:
        id: { type: string, format: uuid }
        request_id: { type: string, format: uuid }
        original_filename: { type: string }
        file_type: { type: string }
        file_size: { type: integer }
        s3_key: { type: string }
        s3_url: { type: string }
        created_at: { type: string, format: date-time }
    Response:
      type: object
      properties:
        id: { type: string, format: uuid }
        request_id: { type: string, format: uuid }
        content: { type: string }
        model: { type: string }
        tokens_used: { type: integer }
        processing_time_ms: { type: integer }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string }
        details: {}
        upload_errors:
          type: array
          items: { type: string }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...